fix unformatted error command response and move event listeneres into separate Cog
